Consolidate domain exports into domain indexes
