+++ src/config/index.js
@@ -1,0 +1,6 @@
+import * as sequelize from './database.js';
+import * as AppError from './error.js';
+import * as logger from './logger.js';
+import * as swaggerJsdoc from './swagger.js';
+
+export { sequelize, AppError, logger, swaggerJsdoc };
